+++ test/Config/SessionConfigTest.php
@@ -9,6 +9,7 @@
 
 namespace ZendTest\Session\Config;
 
+use PHPUnit\Framework\TestCase;
 use Zend\Session\Config\SessionConfig;
 
 /**
@@ -16,7 +17,7 @@ use Zend\Session\Config\SessionConfig;
  * @runTestsInSeparateProcesses
  * @covers Zend\Session\Config\SessionConfig
  */
-class SessionConfigTest extends \PHPUnit_Framework_TestCase
+class SessionConfigTest extends TestCase
 {
     /**
      * @var SessionConfig
@@ -28,19 +29,20 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
         $this->config = new SessionConfig;
     }
 
-    public function assertPhpLessThan71()
+    public function assertPhpLessThan71($message = 'This test requires a PHP version less than 7.1.0')
     {
-        if (version_compare(PHP_VERSION, '7.1.0', '<')) {
+        if (PHP_VERSION_ID < 70100) {
             return true;
         }
-        $this->markTestSkipped('This test requires a PHP version less than 7.1.0');
+        $this->markTestSkipped($message);
     }
 
     // session.save_path
 
     public function testSetSavePathErrorsOnInvalidPath()
     {
-        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException', 'Invalid save_path provided');
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid save_path provided');
         $this->config->setSavePath(__DIR__ . '/foobarboguspath');
     }
 
@@ -65,6 +67,7 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
     {
         $this->config->setPhpSaveHandler('user');
         $this->config->setSavePath('/tmp/sessions.db');
+        $this->assertEquals('/tmp/sessions.db', ini_get('session.save_path'));
     }
 
     // session.name
@@ -111,10 +114,8 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testSettingInvalidSaveHandlerRaisesException()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid save handler specified'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid save handler specified');
         $this->config->setPhpSaveHandler('foobar_bogus');
     }
 
@@ -139,28 +140,22 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testSettingInvalidGcProbabilityRaisesException()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid gc_probability; must be numeric'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid gc_probability; must be numeric');
         $this->config->setGcProbability('foobar_bogus');
     }
 
     public function testSettingInvalidGcProbabilityRaisesException2()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid gc_probability; must be a percentage'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid gc_probability; must be a percentage');
         $this->config->setGcProbability(-1);
     }
 
     public function testSettingInvalidGcProbabilityRaisesException3()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid gc_probability; must be a percentage'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid gc_probability; must be a percentage');
         $this->config->setGcProbability(101);
     }
 
@@ -185,19 +180,15 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testSettingInvalidGcDivisorRaisesException()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid gc_divisor; must be numeric'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid gc_divisor; must be numeric');
         $this->config->setGcDivisor('foobar_bogus');
     }
 
     public function testSettingInvalidGcDivisorRaisesException2()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid gc_divisor; must be a positive integer'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid gc_divisor; must be a positive integer');
         $this->config->setGcDivisor(-1);
     }
 
@@ -222,19 +213,15 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testSettingInvalidGcMaxlifetimeRaisesException()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid gc_maxlifetime; must be numeric'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid gc_maxlifetime; must be numeric');
         $this->config->setGcMaxlifetime('foobar_bogus');
     }
 
     public function testSettingInvalidGcMaxlifetimeRaisesException2()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid gc_maxlifetime; must be a positive integer'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid gc_maxlifetime; must be a positive integer');
         $this->config->setGcMaxlifetime(-1);
     }
 
@@ -261,10 +248,8 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testSettingInvalidSerializeHandlerRaisesException()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid serialize handler specified'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid serialize handler specified');
         $this->config->setSerializeHandler('foobar_bogus');
     }
 
@@ -295,19 +280,15 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testSettingInvalidCookieLifetimeRaisesException()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid cookie_lifetime; must be numeric'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid cookie_lifetime; must be numeric');
         $this->config->setCookieLifetime('foobar_bogus');
     }
 
     public function testSettingInvalidCookieLifetimeRaisesException2()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid cookie_lifetime; must be a positive integer or zero'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid cookie_lifetime; must be a positive integer or zero');
         $this->config->setCookieLifetime(-1);
     }
 
@@ -332,19 +313,22 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testSettingInvalidCookiePathRaisesException()
     {
-        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException', 'Invalid cookie path');
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid cookie path');
         $this->config->setCookiePath(24);
     }
 
     public function testSettingInvalidCookiePathRaisesException2()
     {
-        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException', 'Invalid cookie path');
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid cookie path');
         $this->config->setCookiePath('foo');
     }
 
     public function testSettingInvalidCookiePathRaisesException3()
     {
-        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException', 'Invalid cookie path');
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid cookie path');
         $this->config->setCookiePath('D:\\WINDOWS\\System32\\drivers\\etc\\hosts');
     }
 
@@ -375,19 +359,15 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testSettingInvalidCookieDomainRaisesException()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid cookie domain: must be a string'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid cookie domain: must be a string');
         $this->config->setCookieDomain(24);
     }
 
     public function testSettingInvalidCookieDomainRaisesException2()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'does not match the expected structure for a DNS hostname'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('does not match the expected structure for a DNS hostname');
         $this->config->setCookieDomain('D:\\WINDOWS\\System32\\drivers\\etc\\hosts');
     }
 
@@ -504,78 +484,111 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testSetEntropyFileErrorsOnInvalidPath()
     {
-        $this->assertPhpLessThan71();
-        $this->setExpectedException('Zend\Session\Exception\InvalidArgumentException', 'Invalid entropy_file provided');
+        $this->assertPhpLessThan71('session.entropy_file directive removed in PHP 7.1');
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid entropy_file provided');
         $this->config->setEntropyFile(__DIR__ . '/foobarboguspath');
     }
 
     public function testEntropyFileDefaultsToIniSettings()
     {
-        $this->assertPhpLessThan71();
+        $this->assertPhpLessThan71('session.entropy_file directive removed in PHP 7.1');
         $this->assertSame(ini_get('session.entropy_file'), $this->config->getEntropyFile());
     }
 
     public function testEntropyFileIsMutable()
     {
-        $this->assertPhpLessThan71();
+        $this->assertPhpLessThan71('session.entropy_file directive removed in PHP 7.1');
         $this->config->setEntropyFile(__FILE__);
         $this->assertEquals(__FILE__, $this->config->getEntropyFile());
     }
 
     public function testEntropyFileAltersIniSetting()
     {
-        $this->assertPhpLessThan71();
+        $this->assertPhpLessThan71('session.entropy_file directive removed in PHP 7.1');
         $this->config->setEntropyFile(__FILE__);
         $this->assertEquals(__FILE__, ini_get('session.entropy_file'));
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testSetEntropyFileError()
+    {
+        $this->expectException('PHPUnit\Framework\Error\Deprecated');
+        $this->config->getEntropyFile();
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testGetEntropyFileError()
+    {
+        $this->expectException('PHPUnit\Framework\Error\Deprecated');
+        $this->config->setEntropyFile(__FILE__);
     }
 
     // session.entropy_length
 
     public function testEntropyLengthDefaultsToIniSettings()
     {
-        $this->assertPhpLessThan71();
+        $this->assertPhpLessThan71('session.entropy_length directive removed in PHP 7.1');
         $this->assertSame(ini_get('session.entropy_length'), $this->config->getEntropyLength());
     }
 
     public function testEntropyLengthIsMutable()
     {
-        $this->assertPhpLessThan71();
+        $this->assertPhpLessThan71('session.entropy_length directive removed in PHP 7.1');
         $this->config->setEntropyLength(20);
         $this->assertEquals(20, $this->config->getEntropyLength());
     }
 
     public function testEntropyLengthAltersIniSetting()
     {
-        $this->assertPhpLessThan71();
+        $this->assertPhpLessThan71('session.entropy_length directive removed in PHP 7.1');
         $this->config->setEntropyLength(24);
         $this->assertEquals(24, ini_get('session.entropy_length'));
     }
 
     public function testEntropyLengthCanBeZero()
     {
-        $this->assertPhpLessThan71();
+        $this->assertPhpLessThan71('session.entropy_length directive removed in PHP 7.1');
         $this->config->setEntropyLength(0);
         $this->assertEquals(0, ini_get('session.entropy_length'));
     }
 
     public function testSettingInvalidEntropyLengthRaisesException()
     {
-        $this->assertPhpLessThan71();
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid entropy_length; must be numeric'
-        );
+        $this->assertPhpLessThan71('session.entropy_length directive removed in PHP 7.1');
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid entropy_length; must be numeric');
         $this->config->setEntropyLength('foobar_bogus');
     }
 
     public function testSettingInvalidEntropyLengthRaisesException2()
     {
-        $this->assertPhpLessThan71();
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid entropy_length; must be a positive integer or zero'
-        );
+        $this->assertPhpLessThan71('session.entropy_length directive removed in PHP 7.1');
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid entropy_length; must be a positive integer or zero');
         $this->config->setEntropyLength(-1);
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testGetEntropyLengthError()
+    {
+        $this->expectException('PHPUnit\Framework\Error\Deprecated');
+        $this->config->getEntropyLength();
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testSetEntropyLengthError()
+    {
+        $this->expectException('PHPUnit\Framework\Error\Deprecated');
+        $this->config->setEntropyLength(0);
     }
 
     // session.cache_limiter
@@ -616,10 +629,8 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testSettingInvalidCacheLimiterRaisesException()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid cache limiter provided'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid cache limiter provided');
         $this->config->setCacheLimiter('foobar_bogus');
     }
 
@@ -644,19 +655,15 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testSettingInvalidCacheExpireRaisesException()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid cache_expire; must be numeric'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid cache_expire; must be numeric');
         $this->config->setCacheExpire('foobar_bogus');
     }
 
     public function testSettingInvalidCacheExpireRaisesException2()
     {
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid cache_expire; must be a positive integer'
-        );
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid cache_expire; must be a positive integer');
         $this->config->setCacheExpire(-1);
     }
 
@@ -695,7 +702,7 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testHashFunctionDefaultsToIniSettings()
     {
-        $this->assertPhpLessThan71();
+        $this->assertPhpLessThan71('session.hash_function directive removed in PHP 7.1');
         $this->assertSame(ini_get('session.hash_function'), $this->config->getHashFunction());
     }
 
@@ -704,7 +711,7 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
      */
     public function testHashFunctionIsMutable($hashFunction)
     {
-        $this->assertPhpLessThan71();
+        $this->assertPhpLessThan71('session.hash_function directive removed in PHP 7.1');
         $this->config->setHashFunction($hashFunction);
         $this->assertEquals($hashFunction, $this->config->getHashFunction());
     }
@@ -714,18 +721,34 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
      */
     public function testHashFunctionAltersIniSetting($hashFunction)
     {
-        $this->assertPhpLessThan71();
+        $this->assertPhpLessThan71('session.hash_function directive removed in PHP 7.1');
         $this->config->setHashFunction($hashFunction);
         $this->assertEquals($hashFunction, ini_get('session.hash_function'));
     }
 
     public function testSettingInvalidHashFunctionRaisesException()
     {
-        $this->assertPhpLessThan71();
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid hash function provided'
-        );
+        $this->assertPhpLessThan71('session.hash_function directive removed in PHP 7.1');
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid hash function provided');
+        $this->config->setHashFunction('foobar_bogus');
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testGetHashFunctionError()
+    {
+        $this->expectException('PHPUnit\Framework\Error\Deprecated');
+        $this->config->getHashFunction();
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testSetHashFunctionError()
+    {
+        $this->expectException('PHPUnit\Framework\Error\Deprecated');
         $this->config->setHashFunction('foobar_bogus');
     }
 
@@ -742,6 +765,10 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testHashBitsPerCharacterDefaultsToIniSettings()
     {
+        if (PHP_VERSION_ID >= 70100) {
+            $this->markTestSkipped('session.hash_bits_per_character directive removed in PHP 7.1');
+        }
+
         $this->assertSame(ini_get('session.hash_bits_per_character'), $this->config->getHashBitsPerCharacter());
     }
 
@@ -750,7 +777,7 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
      */
     public function testHashBitsPerCharacterIsMutable($hashBitsPerCharacter)
     {
-        $this->assertPhpLessThan71();
+        $this->assertPhpLessThan71('session.hash_bits_per_character directive removed in PHP 7.1');
         $this->config->setHashBitsPerCharacter($hashBitsPerCharacter);
         $this->assertEquals($hashBitsPerCharacter, $this->config->getHashBitsPerCharacter());
     }
@@ -760,19 +787,142 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
      */
     public function testHashBitsPerCharacterAltersIniSetting($hashBitsPerCharacter)
     {
-        $this->assertPhpLessThan71();
+        $this->assertPhpLessThan71('session.hash_bits_per_character directive removed in PHP 7.1');
         $this->config->setHashBitsPerCharacter($hashBitsPerCharacter);
         $this->assertEquals($hashBitsPerCharacter, ini_get('session.hash_bits_per_character'));
     }
 
     public function testSettingInvalidHashBitsPerCharacterRaisesException()
     {
-        $this->assertPhpLessThan71();
-        $this->setExpectedException(
-            'Zend\Session\Exception\InvalidArgumentException',
-            'Invalid hash bits per character provided'
-        );
+        $this->assertPhpLessThan71('session.hash_bits_per_character directive removed in PHP 7.1');
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid hash bits per character provided');
         $this->config->setHashBitsPerCharacter('foobar_bogus');
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testGetHashBitsPerCharacterError()
+    {
+        $this->expectException('PHPUnit\Framework\Error\Deprecated');
+        $this->config->getHashBitsPerCharacter();
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testSetHashBitsPerCharacterError()
+    {
+        $this->expectException('PHPUnit\Framework\Error\Deprecated');
+        $this->config->setHashBitsPerCharacter(5);
+    }
+
+    // session.sid_length
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testSidLengthDefaultsToIniSettings()
+    {
+        $this->assertSame(ini_get('session.sid_length'), $this->config->getSidLength());
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testSidLengthIsMutable()
+    {
+        $this->config->setSidLength(40);
+        $this->assertEquals(40, $this->config->getSidLength());
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testSidLengthAltersIniSetting()
+    {
+        $this->config->setSidLength(40);
+        $this->assertEquals(40, ini_get('session.sid_length'));
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testSettingInvalidSidLengthRaisesException()
+    {
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid length provided');
+        $this->config->setSidLength('foobar_bogus');
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testSettingOutOfRangeSidLengthRaisesException()
+    {
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid length provided');
+        $this->config->setSidLength(999);
+    }
+
+    // session.sid_bits_per_character
+
+    public function sidSidPerCharacters()
+    {
+        return [
+            [4],
+            [5],
+            [6],
+        ];
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testSidBitsPerCharacterDefaultsToIniSettings()
+    {
+        $this->assertSame(ini_get('session.sid_bits_per_character'), $this->config->getSidBitsPerCharacter());
+    }
+
+    /**
+     * @requires PHP 7.1
+     * @dataProvider sidSidPerCharacters
+     */
+    public function testSidBitsPerCharacterIsMutable($sidBitsPerCharacter)
+    {
+        $this->config->setSidBitsPerCharacter($sidBitsPerCharacter);
+        $this->assertEquals($sidBitsPerCharacter, $this->config->getSidBitsPerCharacter());
+    }
+
+    /**
+     * @requires PHP 7.1
+     * @dataProvider sidSidPerCharacters
+     */
+    public function testSidBitsPerCharacterAltersIniSetting($sidBitsPerCharacter)
+    {
+        $this->config->setSidBitsPerCharacter($sidBitsPerCharacter);
+        $this->assertEquals($sidBitsPerCharacter, ini_get('session.sid_bits_per_character'));
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testSettingInvalidSidBitsPerCharacterRaisesException()
+    {
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid sid bits per character provided');
+        $this->config->setSidBitsPerCharacter('foobar_bogus');
+    }
+
+    /**
+     * @requires PHP 7.1
+     */
+    public function testSettingOutOfBoundSidBitsPerCharacterRaisesException()
+    {
+        $this->expectException('Zend\Session\Exception\InvalidArgumentException');
+        $this->expectExceptionMessage('Invalid sid bits per character provided');
+        $this->config->setSidBitsPerCharacter(999);
     }
 
     // url_rewriter.tags
@@ -815,10 +965,12 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
     public function testSetOptionSetsIniSetting($option, $getter, $value)
     {
         // Leaving out special cases.
-        if ($option != 'remember_me_seconds' && $option != 'url_rewriter_tags') {
-            $this->config->setStorageOption($option, $value);
-            $this->assertEquals(ini_get('session.' . $option), $value);
+        if ($option == 'remember_me_seconds' || $option == 'url_rewriter_tags') {
+            $this->markTestSkipped('remember_me_seconds && url_rewriter_tags');
         }
+
+        $this->config->setStorageOption($option, $value);
+        $this->assertEquals(ini_get('session.' . $option), $value);
     }
 
     public function testSetOptionUrlRewriterTagsGetsMunged()
@@ -830,16 +982,15 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function testSetOptionRememberMeSecondsDoesNothing()
     {
-        // I have no idea how to test this.
+        $this->markTestIncomplete('I have no idea how to test this.');
     }
 
-    /**
-     * @expectedException InvalidArgumentException
-     */
     public function testSetOptionsThrowsExceptionOnInvalidKey()
     {
         $badKey = 'snarfblat';
         $value = 'foobar';
+
+        $this->expectException('InvalidArgumentException');
         $this->config->setStorageOption($badKey, $value);
     }
 
@@ -861,7 +1012,7 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
 
     public function optionsProvider()
     {
-        $options = [
+        $commonOptions = [
             [
                 'save_path',
                 'getSavePath',
@@ -960,35 +1111,49 @@ class SessionConfigTest extends \PHPUnit_Framework_TestCase
         ];
 
         // These options are no longer present as of PHP 7.1.0
-        if (version_compare(PHP_VERSION, '7.1.0', '<')) {
-            $options[] = [
+        if (PHP_VERSION_ID < 70100) {
+            $commonOptions[] = [
                 'entropy_file',
                 'getEntropyFile',
                 __FILE__,
             ];
-            $options[] = [
+
+            $commonOptions[] = [
                 'entropy_length',
                 'getEntropyLength',
                 42,
             ];
-            $options[] = [
-                'hash_bits_per_character',
-                'getOption',
-                5,
-            ];
+
             // The docs do not say this was removed for 7.1.0, but tests fail
             // on 7.1.0 due to this one.
-            $options[] = [
+            $commonOptions[] = [
                 'hash_function',
                 'getHashFunction',
                 'md5',
             ];
+
+            $commonOptions[] = [
+                'hash_bits_per_character',
+                'getOption',
+                5,
+            ];
         }
 
-        // At some point, need to add support for session.sid_bits_per_character
-        if (version_compare(PHP_VERSION, '7.1.0', '>=')) {
+        // New options as of PHP 7.1.0
+        if (PHP_VERSION_ID >= 70100) {
+            $commonOptions[] = [
+                'sid_length',
+                'getSidLength',
+                40,
+            ];
+
+            $commonOptions[] = [
+                'sid_bits_per_character',
+                'getSidBitsPerCharacter',
+                5,
+            ];
         }
 
-        return $options;
+        return $commonOptions;
     }
 }